+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - This was a problem with the Global Chat Detector, Auto Placer, Biome Radar, Ender Energy Distributor, Imbuing Station, and Filteed Platform
 - If the BlockToolModificationEvent is canceled, Slime Blocks will not be turned into Compressed Slime
 - Fixed the Analog Emitter outputting a signal out of its input face (#11)
+- Made Imbue items act more like Potions, in how they return an empty bottle when used (#17)
 
 # 1.3.0